+++ NebulaModel/Packets/Players/PlayerColorChanged.cs
@@ -5,13 +5,13 @@ namespace NebulaModel.Packets.Players
     public class PlayerColorChanged
     {
         public ushort PlayerId { get; set; }
-        public Float3 Color { get; set; }
+        public Float4[] Colors { get; set; }
 
         public PlayerColorChanged() { }
-        public PlayerColorChanged(ushort playerID, Float3 color)
+        public PlayerColorChanged(ushort playerID, Float4[] colors)
         {
             PlayerId = playerID;
-            Color = color;
+            Colors = colors;
         }
     }
 }