+++ NebulaNetwork/PacketProcessors/Session/HandshakeRequestProcessor.cs
@@ -112,7 +112,7 @@ namespace NebulaNetwork.PacketProcessors.Session
             player.Data.Username = !string.IsNullOrWhiteSpace(packet.Username) ? packet.Username : $"Player {player.Id}";
 
             // Add the Mecha Color to the player data
-            player.Data.MechaColor = packet.MechaColor;
+            player.Data.MechaColors = packet.MechaColors;
 
             // Make sure that each player that is currently in the game receives that a new player as join so they can create its RemotePlayerCharacter
             PlayerJoining pdata = new PlayerJoining((PlayerData)player.Data.CreateCopyWithoutMechaData()); // Remove inventory from mecha data